% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{africa_poly}
\alias{africa_poly}
\title{A \code{.poly} file representing the filter file for Africa region.}
\format{
An object of class \code{character} of length 25.
}
\source{
\url{http://download.openstreetmap.fr/polygons/}
}
\usage{
africa_poly
}
\description{
A \code{.poly} file representing the filter file for Africa region.
}
\keyword{datasets}
