% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_poly_to_sfc}
\alias{convert_poly_to_sfc}
\title{Convert .poly file into sfc object}
\usage{
convert_poly_to_sfc(poly)
}
\arguments{
\item{poly}{Character string representing the path of a \code{.poly} file. A URL
can be specified using \code{url()}. See examples.}
}
\value{
A \code{sfc} object.
}
\description{
Convert .poly file into sfc object
}
\details{
The Polygon Filter File Format (\code{.poly}) is defined
\href{https://wiki.openstreetmap.org/wiki/Osmosis/Polygon_Filter_File_Format}{here}.
The code behind the function was inspired by \code{parse_poly} function defined
\href{https://wiki.openstreetmap.org/wiki/Osmosis/Polygon_Filter_File_Python_Parsing}{here}.
}
\examples{
# Starting from text lines stored in an R object
convert_poly_to_sfc(africa_poly)

# Starting from a path pointing to an existing file
africa_path <- system.file("africa.poly", package = "polyread")
convert_poly_to_sfc(africa_path)

# Starting from a connection
\dontrun{
my_url <- "http://download.openstreetmap.fr/polygons/africa.poly"
my_con <- url(my_url)
my_poly <- readLines(my_con)
close(my_con) # don't forget to close the connection
convert_poly_to_sfc(my_poly)
}


}
